requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
pillow>=9.0.0
pyautogui>=0.9.50
keyboard>=0.13.5
python-telegram-bot>=20.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.2
mss>=9.0.0
requests>=2.31.0
pure-python-adb
